fix(agent-vm): extract shared read decisions from lifecycle
Keep agent_vm_lifecycle.py at its line-count baseline by moving the
stale-ready read contract into agent_vm_read.py.

Failure-Class: new
