+++ .github/failure-classes/FC-cached-provider-ready-as-live.json
@@ -4,13 +4,13 @@
   "violated_contract": "A cached Firestore agentVm.status=ready record must not be treated as proof that the corresponding GCE instance is live; status/read paths must distinguish provider NOT_FOUND from API blips and demote or queue fenced reconciler demand instead of leaving eternal ready.",
   "canonical_prevention": "Share one Agent VM read decision helper across desktop status, tools status, and proxy demotion: provider-confirmed NOT_FOUND demotes to updating and queues reconciler demand without wiping ownership; provider/API unknown preserves the owner pointer; ready+running remains unchanged. Cover the three cases with unit tests.",
   "canonical_prevention_artifact": [
-    "backend/services/agent_vm_lifecycle.py",
+    "backend/services/agent_vm_read.py",
     "backend/tests/unit/test_agent_vm_read_decision.py",
     "backend/tests/unit/test_desktop_agent_vm.py"
   ],
   "evidence_prs": [11164],
   "scope_hints": [
-    "backend/services/agent_vm_lifecycle.py",
+    "backend/services/agent_vm_read.py",
     "backend/routers/desktop_agent_vm.py",
     "backend/routers/agent_tools.py",
     "backend/agent-proxy/**"
